+++ instrumentation/net/http/otelhttp/handler_example_test.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,7 +17,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"strings"
@@ -78,7 +77,7 @@ func ExampleNewHandler() {
 					return
 				}
 
-				d, err := ioutil.ReadAll(r.Body)
+				d, err := io.ReadAll(r.Body)
 				if err != nil {
 					log.Println("error reading body: ", err)
 					w.WriteHeader(http.StatusBadRequest)