+++ instrumentation/net/http/otelhttp/handler_test.go
@@ -15,7 +15,7 @@
 package otelhttp_test
 
 import (
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -46,7 +46,7 @@ func TestHandlerReadingNilBodySuccess(t *testing.T) {
 	h := otelhttp.NewHandler(
 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			if r.Body != nil {
-				_, err := ioutil.ReadAll(r.Body)
+				_, err := io.ReadAll(r.Body)
 				assert.NoError(t, err)
 			}
 		}), "test_handler",